Fix writing test files to correct location
