+++ generators/csharp/codegen/src/asIs/test/Pagination/GuidCursorTest.Template.cs
@@ -4,19 +4,20 @@ using <%= namespace%>.Core;
 namespace <%= namespace%>.Test.Core.Pagination;
 
 [TestFixture(Category = "Pagination")]
-public class NoRequestCursorTestCase
+public class GuidCursorTest
 {
     [Test]
-    public async Task CursorPagerShouldWorkWithStringCursor()
+    public async Task CursorPagerShouldWorkWithGuidCursors()
     {
         var pager = CreatePager();
         await AssertPager(pager);
     }
 
-    private const string? Cursor1 = null;
-    private const string Cursor2 = "cursor2";
-    private const string Cursor3 = "cursor3";
-    private string? _cursorCopy;
+    private static readonly Guid? Cursor1 = null;
+    private static readonly Guid Cursor2 = new("00000000-0000-0000-0000-000000000001");
+    private static readonly Guid Cursor3 = new("00000000-0000-0000-0000-000000000001");
+    private Guid? _cursorCopy;
+
     private Pager<object> CreatePager()
     {
         var responses = new List<Response>
@@ -27,7 +28,7 @@ public class NoRequestCursorTestCase
                 {
                     Items = ["item1", "item2"]
                 },
-                Cursor = new ()
+                Cursor = new()
                 {
                     Next = Cursor2
                 }
@@ -38,7 +39,7 @@ public class NoRequestCursorTestCase
                 {
                     Items = ["item1"]
                 },
-                Cursor = new ()
+                Cursor = new()
                 {
                     Next = Cursor3
                 }
@@ -49,7 +50,7 @@ public class NoRequestCursorTestCase
                 {
                     Items = []
                 },
-                Cursor = new ()
+                Cursor = new()
                 {
                     Next = null
                 }
@@ -57,13 +58,16 @@ public class NoRequestCursorTestCase
         }.GetEnumerator();
         _cursorCopy = Cursor1;
         Pager<object> pager = new CursorPager<
-            Request?,
+            Request,
             object?,
             Response,
-            string,
+            Guid?,
             object
         >(
-            null,
+            new()
+            {
+                Cursor = Cursor1
+            },
             null,
             (_, _, _) =>
             {
@@ -109,7 +113,7 @@ public class NoRequestCursorTestCase
 
     private class Request
     {
-        public required string? Cursor { get; set; }
+        public required Guid? Cursor { get; set; }
     }
 
     private class Response
@@ -125,6 +129,6 @@ public class NoRequestCursorTestCase
 
     private class Cursor
     {
-        public required string? Next { get; set; }
+        public required Guid? Next { get; set; }
     }
 }
