+++ generators/csharp/codegen/src/asIs/test/Pagination/IntOffsetTest.Template.cs
@@ -4,7 +4,7 @@ using <%= namespace%>.Core;
 namespace <%= namespace%>.Test.Core.Pagination;
 
 [TestFixture(Category = "Pagination")]
-public class IntOffsetTestCase
+public class IntOffsetTest
 {
     [Test]
     public async Task OffsetPagerShouldWorkWithIntPage()
