+++ generators/csharp/codegen/src/asIs/test/Pagination/NoRequestCursorTest.Template.cs
@@ -1,10 +1,10 @@
 using NUnit.Framework;
-using SeedPagination.Core;
+using <%= namespace%>.Core;
 
-namespace SeedPagination.Test.Core.Pagination;
+namespace <%= namespace%>.Test.Core.Pagination;
 
 [TestFixture(Category = "Pagination")]
-public class NoRequestCursorTestCase
+public class NoRequestCursorTest
 {
     [Test]
     public async Task CursorPagerShouldWorkWithStringCursor()
@@ -17,29 +17,52 @@ public class NoRequestCursorTestCase
     private const string Cursor2 = "cursor2";
     private const string Cursor3 = "cursor3";
     private string? _cursorCopy;
-
     private Pager<object> CreatePager()
     {
         var responses = new List<Response>
         {
             new()
             {
-                Data = new() { Items = ["item1", "item2"] },
-                Cursor = new() { Next = Cursor2 },
+                Data = new()
+                {
+                    Items = ["item1", "item2"]
+                },
+                Cursor = new ()
+                {
+                    Next = Cursor2
+                }
             },
             new()
             {
-                Data = new() { Items = ["item1"] },
-                Cursor = new() { Next = Cursor3 },
+                Data = new()
+                {
+                    Items = ["item1"]
+                },
+                Cursor = new ()
+                {
+                    Next = Cursor3
+                }
             },
             new()
             {
-                Data = new() { Items = [] },
-                Cursor = new() { Next = null },
-            },
+                Data = new()
+                {
+                    Items = []
+                },
+                Cursor = new ()
+                {
+                    Next = null
+                }
+            }
         }.GetEnumerator();
         _cursorCopy = Cursor1;
-        Pager<object> pager = new CursorPager<Request?, object?, Response, string, object>(
+        Pager<object> pager = new CursorPager<
+            Request?,
+            object?,
+            Response,
+            string,
+            object
+        >(
             null,
             null,
             (_, _, _) =>
