+++ generators/csharp/codegen/src/asIs/test/Pagination/StepOffsetTest.Template.cs
@@ -4,7 +4,7 @@ using <%= namespace%>.Core;
 namespace <%= namespace%>.Test.Core.Pagination;
 
 [TestFixture(Category = "Pagination")]
-public class StepPageOffsetPaginationTestCase
+public class StepPageOffsetPaginationTest
 {
     [Test]
     public async Task OffsetPagerShouldWorkWithStep()
